+++ internal/model/address.go
@@ -6,8 +6,8 @@ import (
 
 // Address represents a shipping address
 type Address struct {
-	ID         uint      `json:"id" gorm:"column:id;primaryKey"`
-	UserID     uint      `json:"user_id" gorm:"column:user_id;index"`
+	ID         uint64    `json:"id" gorm:"column:id;primaryKey"`
+	UserID     uint64    `json:"user_id" gorm:"column:user_id;index"`
 	Name       string    `json:"name" gorm:"column:name;not null"`
 	Phone      string    `json:"phone" gorm:"column:phone;not null"`
 	Province   string    `json:"province" gorm:"column:province;not null"`
